requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
tensorflow
google-generativeai